+++ examples/engine_cross_compilation.rs
@@ -19,8 +19,7 @@
 //! Ready?
 
 use std::str::FromStr;
-use wasmer::{wat2wasm, Module, RuntimeError, Store};
-use wasmer_compiler::EngineBuilder;
+use wasmer::{wat2wasm, EngineBuilder, Module, RuntimeError, Store};
 use wasmer_compiler_cranelift::Cranelift;
 use wasmer_types::{CpuFeature, Target, Triple};
 
@@ -70,10 +69,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // That's where we specify the target for the compiler.
     //
     // Use the Universal engine.
-    let mut engine = EngineBuilder::new(compiler_config, Some(target), None).engine();
+    let mut engine = EngineBuilder::new(compiler_config).set_target(Some(target));
 
     // Create a store, that holds the engine.
-    let mut store = Store::new_with_engine(&engine);
+    let mut store = Store::new(engine);
 
     println!("Compiling module...");
     // Let's compile the Wasm module.