+++ examples/instance.rs
@@ -15,7 +15,6 @@
 //! Ready?
 
 use wasmer::{imports, wat2wasm, FunctionEnv, Instance, Module, Store, TypedFunction};
-use wasmer_compiler::EngineBuilder;
 use wasmer_compiler_cranelift::Cranelift;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -39,8 +38,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Note that we don't need to specify the engine/compiler if we want to use
     // the default provided by Wasmer.
     // You can use `Store::default()` for that.
-    let mut store =
-        Store::new_with_engine(&EngineBuilder::new(Cranelift::default(), None, None).engine());
+    let mut store = Store::new(Cranelift::default());
     let mut env = FunctionEnv::new(&mut store, ());
 
     println!("Compiling module...");