+++ examples/memory.rs
@@ -18,7 +18,6 @@ use std::mem;
 use wasmer::{
     imports, wat2wasm, Bytes, FunctionEnv, Instance, Module, Pages, Store, TypedFunction,
 };
-use wasmer_compiler::EngineBuilder;
 use wasmer_compiler_cranelift::Cranelift;
 
 // this example is a work in progress:
@@ -59,8 +58,7 @@ fn main() -> anyhow::Result<()> {
     // Note that we don't need to specify the engine/compiler if we want to use
     // the default provided by Wasmer.
     // You can use `Store::default()` for that.
-    let mut store =
-        Store::new_with_engine(&EngineBuilder::new(Cranelift::default(), None, None).engine());
+    let mut store = Store::new(Cranelift::default());
     let mut env = FunctionEnv::new(&mut store, ());
 
     println!("Compiling module...");